requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
filelock==3.12.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
patchify==0.2.3
Pillow==9.5.0
requests==2.29.0
sympy==1.11.1
torch==2.0.0
torchaudio==2.0.1
torchgeometry==0.1.2
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
